Task4 - AZURE LANGUAGE SERVICES/Recognize linked entities - Task4/requirement.txt
azure-ai-textanalytics
azure-core==1.23.0
Django==3.2.6